success! only one filtered data emit
